+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 * Update tests and require latest version of pylint (>=1.8), fixes
   [#53](https://github.com/landscapeio/pylint-django/issues/53),
   [#97](https://github.com/landscapeio/pylint-django/issues/97)
+* [#81](https://github.com/landscapeio/pylint-django/issues/81) Fix 'duplicate-except' false negative
+  for except blocks which catch the `DoesNotExist` exception.
 
 ## Version 0.7.4
 * [#88](https://github.com/landscapeio/pylint-django/pull/88) Fixed builds with Django 1.10 (thanks to [federicobond](https://github.com/federicobond))